Moved colors.js file to a global level
